python/requirements.txt
requests
python-dotenv
rich
chromadb
sentence-transformers
pyyaml
boto3
